Every .swift file had a change here. Should have split this up.
UIAppearance is, like, the greatest thing ever.
replaced UnreadCounter's circle with a triangle, moved appearances of it to the top right of the superview
cleaned up how background refresh is done.
Got the swipe to show actions working on the tableviews.
Made the FindFeedViewController more generic - able to simply be a web-browser
Switched article cell back to using textviews, with correctly rendered html!
ArticleCells now only show either the summary or nothing as a preview.
Changed how articles can be viewed - defaults to showing either the content or the summary, with a button to toggle between that and the link, or just loading the linked article, if a content or summary is unavailable.
Turned on background refresh.