badges the icon.
OPMLParser now is actually case-insensitive, but only where it matters (some URLS are case-sensitive)
OPMLParser now imports things other than the url for a feed (tags, and it also imports query feeds!)

Added support for renaming a feed by giving it a tag prefixed with '~', the '~' is removed, and the rest of the tag is displayed to the user as the feed's name. The only way to add tags to something is to import an OPML file that has tag data.

DataManager now creates OPML files which are more human-readable.

Fixed a crash in FindFeedViewController if you have query feeds.
Fixed bug displaying local import controller.