Re-implemented QueryFeedViewController as a subclass of UITableViewController.
Fleshed out TagPickerView, so that it's more than just an idea.
FeedViewController and QueryFeedViewController will auto-save for you.
Created TagEditorViewController, which does exactly what it says on the tin (allows you to create and edit tags for feeds).
Loading bar in ArticleViewController now works.

Added a "flags" property to Articles, with appropriate extension to get them as a list of Strings.
Removed unused "preview" property of Articles.