Use Messenger for when opml exporting produces an error.
Justification for this error is that the user can just try again.
However, more than likely, they'll hit the exact same error anyway.
The likely cause of errors are: The database getting fucked somehow, and the filesystem being full.
The first is highly unlikely, and the second isn't really actionable on our part.

[#162826824]